Add let user delete courses from the courses page
Added route, controller logic, query logic, and buttons on the courses
page to delete a selected course. This also necessitated creating a new
JS script for the html page that selects each delete button and adds an
event listener with the requesit logic. Axios is used, but due to
difficulties in importing it with both ES6 modules and CommonJS, an CDN
is used (for now).