Add functionality to let user update course name
Updated views, routes, controllers and queries to allow put request of
this kind. Some notes:

- The user first clicks on an update button on the courses page and is
redirected another page to enter the update information. Since the
update is done by course id, it was necessary to send this info from
the update page to the updateCourseName controller with a post request,
and this was done by passing the template the id into a form element
with display: none. This feels a little hacky, so look into better ways
to do this.
- Since there were issues with getting the updatecours.ejs to read the
script.js in the /public dir, the axios put request was implemented in a
script tag at the bottom of the page. This is also less than ideal, so
it would be best to find a way to move this into the /public/script.js
file.